SAC-API/requirements.txt
flask
python-dotenv
openai
faiss-cpu
numpy
tqdm
requests
